requirements.txt
absl-py==2.3.1
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
grpcio==1.74.0
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.7.1
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.14.1
Werkzeug==3.1.3
